% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codons.R
\name{get_codons}
\alias{get_codons}
\title{get_codons}
\usage{
get_codons(input_sequence, start = 1)
}
\arguments{
\item{input_sequence}{Input RNA sequence}

\item{start}{Starting index}
}
\value{
codons as 3 character substrings of the input sequence
}
\description{
get_codons
}
\examples{
example_sequence <- "ACUGAGUCCAGUUCAAAA"
get_codons(example_sequence)

get_codons(example_sequence, start = 4)
}
