% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_seq.R
\name{make_random_seq}
\alias{make_random_seq}
\title{Make random sequence}
\usage{
make_random_seq(size)
}
\arguments{
\item{size}{this is the length of your random sequence}
}
\value{
return a specified size of random DNA sequence
}
\description{
Make random sequence
}
\examples{
make_random_seq(10)
}
