% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotaminoacids.R
\name{plot_amino_acids}
\alias{plot_amino_acids}
\title{Amino acid bar plot}
\usage{
plot_amino_acids(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{amino acid sequence string}
}
\value{
Plot of amino acid counts
}
\description{
Amino acid bar plot
}
\examples{
example_aa_seq <- "FVVSQAGL"
plot_amino_acids(example_aa_seq)
}
