% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript.R
\name{trans_cription}
\alias{trans_cription}
\title{Transcription}
\usage{
trans_cription(template_seq)
}
\arguments{
\item{template_seq}{the function substitute T to U}
}
\value{
transcript of the template
}
\description{
Transcription
}
