% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_codons.R
\name{translate_codons}
\alias{translate_codons}
\title{translate codons}
\usage{
translate_codons(codon_input)
}
\arguments{
\item{codon_input}{}
}
\value{
concatenated codon object translated to amino acids in string format
}
\description{
translate codons
}
\examples{
example_vector <- c("AUG", "ACU", "GAG", "UCC", "AGU", "UCA", "AAA")
translate_codons(example_vector)
}
